+++ src/main/java/net/geforcemods/securitycraft/api/IOwnable.java
@@ -80,16 +80,16 @@ public interface IOwnable {
 	}
 
 	/**
-	 * Checks whether the given player owns this IOwnable.
+	 * Checks whether the given entity owns this IOwnable.
 	 *
-	 * @param player The player to check ownership of
-	 * @return true if the given player owns this IOwnable, false otherwise
+	 * @param entity The entity to check ownership of
+	 * @return true if the given entity owns this IOwnable, false otherwise
 	 */
-	public default boolean isOwnedBy(PlayerEntity player) {
-		if (player == null)
+	public default boolean isOwnedBy(Entity entity) {
+		if (entity instanceof PlayerEntity)
+			return isOwnedBy(new Owner((PlayerEntity) entity));
+		else
 			return false;
-
-		return isOwnedBy(new Owner(player));
 	}
 
 	/**