+++ src/main/java/net/geforcemods/securitycraft/blockentities/ProtectoBlockEntity.java
@@ -9,8 +9,8 @@ import net.geforcemods.securitycraft.api.Option.IgnoreOwnerOption;
 import net.geforcemods.securitycraft.blocks.ProtectoBlock;
 import net.geforcemods.securitycraft.entity.sentry.Sentry;
 import net.geforcemods.securitycraft.misc.ModuleType;
-import net.geforcemods.securitycraft.util.EntityUtils;
 import net.geforcemods.securitycraft.util.LevelUtils;
+import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
@@ -45,7 +45,7 @@ public class ProtectoBlockEntity extends DisguisableBlockEntity implements ITick
 				boolean shouldDeactivate = false;
 
 				for (LivingEntity entity : entities) {
-					if (!(entity instanceof Sentry) && !EntityUtils.isInvisible(entity)) {
+					if (!(entity instanceof Sentry) && !Utils.isEntityInvisible(entity)) {
 						if (entity instanceof PlayerEntity) {
 							PlayerEntity player = (PlayerEntity) entity;