+++ src/main/java/net/geforcemods/securitycraft/blockentities/RetinalScannerBlockEntity.java
@@ -26,7 +26,6 @@ import net.geforcemods.securitycraft.api.Owner;
 import net.geforcemods.securitycraft.blocks.RetinalScannerBlock;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.util.BlockUtils;
-import net.geforcemods.securitycraft.util.EntityUtils;
 import net.geforcemods.securitycraft.util.PlayerUtils;
 import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.block.Block;
@@ -82,7 +81,7 @@ public class RetinalScannerBlockEntity extends DisguisableBlockEntity implements
 
 			int signalLength = getSignalLength();
 
-			if ((!state.getValue(RetinalScannerBlock.POWERED) || signalLength == 0) && !EntityUtils.isInvisible(entity)) {
+			if ((!state.getValue(RetinalScannerBlock.POWERED) || signalLength == 0) && !Utils.isEntityInvisible(entity)) {
 				if (entity instanceof PlayerEntity) {
 					PlayerEntity player = (PlayerEntity) entity;
 					Owner viewingPlayer = new Owner(player);