+++ src/main/java/net/geforcemods/securitycraft/blockentities/ScannerTrapdoorBlockEntity.java
@@ -13,7 +13,6 @@ import net.geforcemods.securitycraft.api.Option.IntOption;
 import net.geforcemods.securitycraft.api.Owner;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.util.BlockUtils;
-import net.geforcemods.securitycraft.util.EntityUtils;
 import net.geforcemods.securitycraft.util.PlayerUtils;
 import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.block.Block;
@@ -53,7 +52,7 @@ public class ScannerTrapdoorBlockEntity extends CustomizableBlockEntity implemen
 
 	@Override
 	public boolean onEntityViewed(LivingEntity entity, BlockRayTraceResult hitResult) {
-		if (!EntityUtils.isInvisible(entity)) {
+		if (!Utils.isEntityInvisible(entity)) {
 			BlockState state = getBlockState();
 
 			if (!(entity instanceof PlayerEntity) || !(state.getValue(TrapDoorBlock.OPEN) ? hitResult.getDirection().getAxis() == state.getValue(HorizontalBlock.FACING).getAxis() : hitResult.getDirection().getAxis() == Axis.Y))